move serde types into a nexus-free core crate
pure move, the addon re-exports everything from the old paths. point is
cargo test -p event_timers_core now runs natively instead of needing
wine or a windows box.
